Add managed cluster auto-approval for awsirsa
